split advanced controls up, fix coderabbit suggestion
