+++ apps/harvester/src/Advanced/types.ts
@@ -1,0 +1,2 @@
+
+export type DimmerCallback = (message: string, callback: () => Promise<void>) => Promise<void>;
